.claude/skills/document-ocr/requirements.txt
google-genai
python-dotenv
pdf2image
Pillow
pillow-heif
